@@ -3,51 +3,52 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@amcharts/amcharts4": "^4.10.22",
+    "@amcharts/amcharts4": "^4.10.25",
     "@amcharts/amcharts4-fonts": "^4.0.1",
-    "@amcharts/amcharts4-geodata": "^4.1.22",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.60",
+    "@amcharts/amcharts4-geodata": "^4.1.23",
+    "@material-ui/core": "^4.12.4",
+    "@material-ui/icons": "^4.11.3",
+    "@material-ui/lab": "4.0.0-alpha.61",
     "@material-ui/pickers": "^3.3.10",
-    "@material-ui/styles": "^4.11.4",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.4.1",
-    "axios": "^0.23.0",
+    "@material-ui/styles": "^4.11.5",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.1.1",
+    "@testing-library/user-event": "^14.1.1",
+    "axios": "^0.27.2",
     "clsx": "^1.1.1",
-    "core-js": "^3.18.3",
-    "date-fns": "^2.25.0",
+    "core-js": "^3.22.3",
+    "date-fns": "^2.28.0",
     "draft-js": "^0.11.7",
     "draftjs-to-html": "^0.9.1",
-    "notistack": "^1.0.10",
-    "react": "^17.0.2",
+    "node-sass": "^7.0.1",
+    "notistack": "^2.0.4",
+    "react": "^18.1.0",
     "react-custom-scrollbars": "^4.2.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.1.0",
     "react-draft-wysiwyg": "^1.14.7",
     "react-icons": "^4.3.1",
     "react-infinite-scroll-component": "^6.1.0",
-    "react-number-format": "^4.7.3",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "^4.0.3",
+    "react-number-format": "^4.9.3",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
     "regenerator-runtime": "^0.13.9",
-    "socket.io-client": "4.3.0",
-    "styled-components": "^5.3.1",
-    "sweetalert2": "^11.1.8",
-    "typescript": "^4.4.4",
-    "web-vitals": "^2.1.2",
-    "workbox-background-sync": "^6.3.0",
-    "workbox-broadcast-update": "^6.3.0",
-    "workbox-cacheable-response": "^6.3.0",
-    "workbox-core": "^6.3.0",
-    "workbox-expiration": "^6.3.0",
-    "workbox-google-analytics": "^6.3.0",
-    "workbox-navigation-preload": "^6.3.0",
-    "workbox-precaching": "^6.3.0",
-    "workbox-range-requests": "^6.3.0",
-    "workbox-routing": "^6.3.0",
-    "workbox-strategies": "^6.3.0",
-    "workbox-streams": "^6.3.0"
+    "socket.io-client": "4.5.0",
+    "styled-components": "^5.3.5",
+    "sweetalert2": "^11.4.9",
+    "typescript": "^4.6.4",
+    "web-vitals": "^2.1.4",
+    "workbox-background-sync": "^6.5.3",
+    "workbox-broadcast-update": "^6.5.3",
+    "workbox-cacheable-response": "^6.5.3",
+    "workbox-core": "^6.5.3",
+    "workbox-expiration": "^6.5.3",
+    "workbox-google-analytics": "^6.5.3",
+    "workbox-navigation-preload": "^6.5.3",
+    "workbox-precaching": "^6.5.3",
+    "workbox-range-requests": "^6.5.3",
+    "workbox-routing": "^6.5.3",
+    "workbox-strategies": "^6.5.3",
+    "workbox-streams": "^6.5.3"
   },
   "scripts": {
     "eslint": "eslint ./src",
@@ -79,28 +80,28 @@
   "devDependencies": {
     "@types/core-js": "^2.5.5",
     "@types/draftjs-to-html": "^0.8.1",
-    "@types/jest": "^27.0.2",
-    "@types/lodash": "^4.14.175",
-    "@types/node": "^16.11.0",
-    "@types/react": "^17.0.30",
-    "@types/react-custom-scrollbars": "^4.0.8",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-draft-wysiwyg": "^1.13.3",
-    "@types/react-router-dom": "^5.3.2",
+    "@types/jest": "^27.4.1",
+    "@types/lodash": "^4.14.182",
+    "@types/node": "^17.0.29",
+    "@types/react": "^18.0.8",
+    "@types/react-custom-scrollbars": "^4.0.10",
+    "@types/react-dom": "^18.0.0",
+    "@types/react-draft-wysiwyg": "^1.13.4",
+    "@types/react-router-dom": "^5.3.3",
     "@types/regenerator-runtime": "^0.13.1",
-    "@types/styled-components": "^5.1.15",
-    "@typescript-eslint/eslint-plugin": "^5.0.0",
-    "@typescript-eslint/parser": "^5.0.0",
-    "eslint": "^7.11.0",
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-import-resolver-typescript": "^2.5.0",
-    "eslint-plugin-import": "^2.25.2",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
+    "@types/styled-components": "^5.1.25",
+    "@typescript-eslint/eslint-plugin": "^5.21.0",
+    "@typescript-eslint/parser": "^5.21.0",
+    "eslint": "^8.14.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-import-resolver-typescript": "^2.7.1",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.26.1",
-    "husky": "^7.0.2",
+    "eslint-plugin-react": "^7.29.4",
+    "husky": "^7.0.4",
     "inquirer-directory": "^2.2.0",
-    "prettier": "^2.4.1"
+    "prettier": "^2.6.2"
   }
 }
