@@ -19,7 +19,6 @@ module.exports = {
   },
   plugins: ['react', 'import', 'jsx-a11y', 'prettier'],
   rules: {
-    'no-confusing-arrow': 'off',
     'max-len': [
       'error',
       {
@@ -34,37 +33,7 @@ module.exports = {
       },
     ],
     camelcase: 'off',
-    'react/react-in-jsx-scope': 'off',
-    '@typescript-eslint/camelcase': 'off',
-    'no-param-reassign': 'off',
-    'react/jsx-curly-newline': 'off',
-    'react/jsx-wrap-multilines': 'off',
-    'operator-linebreak': 'off',
-    'object-curly-newline': 'off',
-    'function-paren-newline': 'off',
-    'react/no-unused-prop-types': 'off',
-    'react/button-has-type': 'off',
-    'implicit-arrow-linebreak': 'off',
-    'react/jsx-one-expression-per-line': 'off',
-    'comma-dangle': 'off',
-    'no-nested-ternary': 'off',
-    'import/no-extraneous-dependencies': 'off',
-    '@typescript-eslint/explicit-module-boundary-types': 'off',
-    semi: 'off',
-    'react-hooks/exhaustive-deps': 'off',
-    '@typescript-eslint/no-explicit-any': 'off',
-    'react/require-default-props': 'off',
-    'react/jsx-props-no-spreading': 'off',
-    'no-use-before-define': 'off',
-    'react/jsx-filename-extension': [
-      'error',
-      {
-        extensions: ['.tsx', '.ts', '.js'],
-      },
-    ],
     'import/prefer-default-export': 'off',
-    '@typescript-eslint/explicit-function-return-type': 'off',
-    '@typescript-eslint/explicit-member-accessibility': 'off',
     'import/extensions': [
       'error',
       'ignorePackages',
@@ -75,6 +44,39 @@ module.exports = {
         tsx: 'never',
       },
     ],
+    'react/jsx-filename-extension': [
+      'error',
+      {
+        extensions: ['.tsx', '.ts', '.js'],
+      },
+    ],
+    'react/react-in-jsx-scope': 'off',
+    'prefer-regex-literals': 'off',
+    'no-use-before-define': 'off',
+    'react/function-component-definition': 'off',
+    // 'no-confusing-arrow': 'off',
+    // '@typescript-eslint/camelcase': 'off',
+    // 'no-param-reassign': 'off',
+    // 'react/jsx-curly-newline': 'off',
+    // 'react/jsx-wrap-multilines': 'off',
+    // 'operator-linebreak': 'off',
+    // 'object-curly-newline': 'off',
+    // 'function-paren-newline': 'off',
+    // 'react/no-unused-prop-types': 'off',
+    // 'react/button-has-type': 'off',
+    // 'implicit-arrow-linebreak': 'off',
+    // 'react/jsx-one-expression-per-line': 'off',
+    // 'comma-dangle': 'off',
+    // 'no-nested-ternary': 'off',
+    // 'import/no-extraneous-dependencies': 'off',
+    // '@typescript-eslint/explicit-module-boundary-types': 'off',
+    // semi: 'off',
+    // 'react-hooks/exhaustive-deps': 'off',
+    // '@typescript-eslint/no-explicit-any': 'off',
+    // 'react/require-default-props': 'off',
+    // 'react/jsx-props-no-spreading': 'off',
+    // '@typescript-eslint/explicit-function-return-type': 'off',
+    // '@typescript-eslint/explicit-member-accessibility': 'off',
   },
   settings: {
     'import/parsers': {
