@@ -1,10 +1,12 @@
 import GlobalStyle from '../GlobalStyle';
 
-export const App = () => (
-  <>
-    <GlobalStyle />
-    <h1>Boilerplate ReactJS</h1>
-  </>
-);
+export function App() {
+  return (
+    <>
+      <GlobalStyle />
+      <h1>Boilerplate ReactJS</h1>
+    </>
+  );
+}
 
 export default App;
